requirements.txt
langgraph
langchain
langchain-community
openai
transformers
torch
pydantic
python-dotenv
langchain_openai
beautifulsoup4
langchain-tavily
langgraph-supervisor